+++ server/src/api/controllers/interfaces/IAdminReqBody.ts
@@ -1,0 +1,6 @@
+interface IAdminReqBody {
+  name: string;
+  password: string;
+}
+
+export default IAdminReqBody;